requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
kafka-python==2.0.2
py4j==0.10.9.7
pyspark==3.5.3
requests==2.32.3
urllib3==2.2.3
psycopg2-binary
